move Summary model to comment package
