+++ src/common/entities/summary.py
