+++ src/six_fourty_nine/entities/prize_breakdown.py
@@ -2,7 +2,7 @@
 from pydantic import BaseModel, Field
 
 from src.common.entities.numbers_matched import NumbersMatched
-from .summary import Summary
+from src.common.entities.summary import Summary
 
 class PrizeBreakdown(BaseModel):
     """Model represent prize statistic for 6/49 draw"""
